Manage yproxy connection lifetime with a shared RAII scope guard
Also move auxuliar class (ScopeGuard) to its own file.
